+++ Engine/src/Pipelines/GraphicsPipeline.cpp
@@ -54,9 +54,9 @@ void GraphicsPipeline::createPipelineLayout(const GraphicsPipelineConfiguration&
     
     VkPipelineLayoutCreateInfo pipelineLayoutInfo{};
     pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-    pipelineLayoutInfo.setLayoutCount = config.shader->getDescriptorSetLayouts().size();
+    pipelineLayoutInfo.setLayoutCount = static_cast<uint32_t>(config.shader->getDescriptorSetLayouts().size());
     pipelineLayoutInfo.pSetLayouts = config.shader->getDescriptorSetLayouts().data();
-    pipelineLayoutInfo.pushConstantRangeCount = config.shader->getPushConstantLayouts().size();
+    pipelineLayoutInfo.pushConstantRangeCount = static_cast<uint32_t>(config.shader->getPushConstantLayouts().size());
     pipelineLayoutInfo.pPushConstantRanges = config.shader->getPushConstantLayouts().data();
 
     if (vkCreatePipelineLayout(device, &pipelineLayoutInfo, nullptr, &m_pipelineLayout) != VK_SUCCESS) {
@@ -74,7 +74,7 @@ void GraphicsPipeline::createPipeline(const GraphicsPipelineConfiguration& confi
 
     VkPipelineRenderingCreateInfoKHR pipelineDynamicRenderingInfo{};
     pipelineDynamicRenderingInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_RENDERING_CREATE_INFO_KHR;
-    pipelineDynamicRenderingInfo.colorAttachmentCount = config.framebufferSpec.colorAttachments.size();
+    pipelineDynamicRenderingInfo.colorAttachmentCount = static_cast<uint32_t>(config.framebufferSpec.colorAttachments.size());
     pipelineDynamicRenderingInfo.pColorAttachmentFormats = config.framebufferSpec.colorAttachments.data();
     pipelineDynamicRenderingInfo.depthAttachmentFormat = config.framebufferSpec.depthAttachment;
     pipelineDynamicRenderingInfo.stencilAttachmentFormat = config.framebufferSpec.stencilAttachment;
@@ -82,7 +82,7 @@ void GraphicsPipeline::createPipeline(const GraphicsPipelineConfiguration& confi
 
     VkGraphicsPipelineCreateInfo pipelineInfo{};
     pipelineInfo.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-    pipelineInfo.stageCount = config.shader->getStages().size();
+    pipelineInfo.stageCount = static_cast<uint32_t>(config.shader->getStages().size());
     pipelineInfo.pStages = config.shader->getStages().data();
 
     pipelineInfo.pVertexInputState = &config.vertexInputState;
